[FEATURE] Added new auto write code
New auto writing system in place.  Create the auto write from the
Pipeline Tools menu and once you have this node in, it will continue to
update based on your changes in the script with resolution and
colorspace, etc.  You can also change the file type with the drop down
and the path will update automatically also.